worker/requirements.txt
asyncpg>=0.29.0
httpx>=0.27.0
mlx>=0.21.0
mlx-lm>=0.21.0
mlx-vlm>=0.1.0
playwright>=1.48.0
python-dotenv>=1.0.0
